Script is ready, but won't be merged. Waiting clarification of usage common functions.
