+++ test_scripts/Polices/appID_Management/163_ATF_P_TC_HMI_Status_Value_Of_AppId_In_PT_Is_Null.lua
@@ -1,0 +1,91 @@
+---------------------------------------------------------------------------------------------
+-- Requirement summary:
+-- [HMI Status]: The value of <appID> in LocalPT is null
+--
+-- Description:
+-- SDL should only allow an HMILevel of NONE to the app
+-- in case PolicyTable has "<appID>": "null" in the Local PolicyTable
+--
+-- Preconditions:
+-- 1. appID="123_xyz" is registered to SDL yet
+-- 2. appID="123_xyz" has null policies
+-- Steps:
+-- 1. Activate app
+-- 2. Verify status of activation
+-- 3. Verify app hmi level
+--
+-- Expected result:
+-- 1. status = "REJECTED"
+-- 2. hmiLevel = "NONE"
+---------------------------------------------------------------------------------------------
+
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+
+--[[ Required Shared libraries ]]
+local mobileSession = require("mobile_session")
+local testCasesForPolicyAppIdManagament = require("user_modules/shared_testcases/testCasesForPolicyAppIdManagament")
+local commonFunctions = require("user_modules/shared_testcases/commonFunctions")
+local commonSteps = require("user_modules/shared_testcases/commonSteps")
+local testCasesForPolicyTable = require('user_modules/shared_testcases/testCasesForPolicyTable')
+
+--[[ General Precondition before ATF start ]]
+commonSteps:DeleteLogsFileAndPolicyTable()
+
+--[[ General Settings for configuration ]]
+Test = require("connecttest")
+require("user_modules/AppTypes")
+
+local HMIAppID
+--[[ Preconditions ]]
+commonFunctions:newTestCasesGroup("Preconditions")
+function Test:Pecondition_trigger_getting_device_consent()
+  testCasesForPolicyTable:trigger_getting_device_consent(self, config.application1.registerAppInterfaceParams.appName, config.deviceMAC)
+end
+
+function Test:UpdatePolicy()
+  testCasesForPolicyAppIdManagament:updatePolicyTable(self, "files/jsons/Policies/appID_Management/ptu_013_1.json")
+end
+
+function Test:StartNewSession()
+  self.mobileSession2 = mobileSession.MobileSession(self, self.mobileConnection)
+  self.mobileSession2:StartService(7)
+end
+
+function Test:RegisterNewApp()
+  config.application2.registerAppInterfaceParams.appName = "Media Application"
+  config.application2.registerAppInterfaceParams.appID = "123_xyz"
+  local correlationId = self.mobileSession2:SendRPC("RegisterAppInterface", config.application2.registerAppInterfaceParams)
+  EXPECT_HMINOTIFICATION("BasicCommunication.OnAppRegistered")
+  :Do(function(_,data)
+      HMIAppID = data.params.application.appID
+      self.applications[config.application2.registerAppInterfaceParams.appName] = data.params.application.appID
+    end)
+
+  self.mobileSession2:ExpectNotification("OnHMIStatus", {hmiLevel = "NONE", systemContext = "MAIN", audioStreamingState = "NOT_AUDIBLE" })
+  self.mobileSession2:ExpectResponse(correlationId, { success = true })
+end
+
+function Test:ActivateApp()
+  local requestId = self.hmiConnection:SendRequest("SDL.ActivateApp", { appID = HMIAppID})
+  -- code = 0: SUCCESS
+  EXPECT_HMIRESPONSE(requestId,{result = {code = 0, method = "SDL.ActivateApp"}})
+  self.mobileSession2:ExpectNotification("OnHMIStatus", {hmiLevel = "FULL", systemContext = "MAIN"})
+end
+
+function Test:Precondition_UpdatePolicy()
+  EXPECT_HMICALL("BasicCommunication.PolicyUpdate")
+  testCasesForPolicyAppIdManagament:updatePolicyTable(self, "files/jsons/Policies/appID_Management/ptu_013_2.json")
+  self.mobileSession2:ExpectNotification("OnHMIStatus", { hmiLevel ="NONE", systemContext = "MAIN", audioStreamingState = "NOT_AUDIBLE" })
+end
+
+--[[ Test ]]
+commonFunctions:newTestCasesGroup("Test")
+function Test:TestStep_ActivateApp()
+  local requestId = self.hmiConnection:SendRequest("SDL.ActivateApp", { appID = HMIAppID})
+  -- code = 4: REJECTED
+  EXPECT_HMIRESPONSE(requestId,{result = {code = 4, method = "SDL.ActivateApp"}})
+  self.mobileSession2:ExpectNotification("OnHMIStatus", { hmiLevel = "NONE" })
+end
+
+return Test
